install: Only set affinity for Cilium agent
This commit is only intended for the v1.7 tree.

This fixes an issue in v1.7 CI where the ManagedEtcd tests failed due to
a unscheduleable cilium-etcd-operator and the cilium agent, causing the
tests to hang. The reason is that b125bd15c assumed that only the Cilium
agent would pick up the anti-affinity changes. Because the
cilium-etcd-operator Chart also references `Values.global.affinity`, its
anti-affinity rules were erroneously applied. There are certain tests
which configure cilium-etcd-operator by adding `global.affinity` in the
Helm template, so the fix is to reduce the scope of the anti-affinity
rules to just the Cilium agent Chart.

Fixes: b125bd15c ("Add anti-affinity for Cilium pods")

Signed-off-by: Chris Tarazi <chris@isovalent.com>